Remove left and right arrow
In decided to restrict the number of digits to keep the feel of
and old-school calculator with a limited display. I will also
limit the output, so that if a number exceeds the amount of digits
in the display, the calculator displays and error.

I replaced the arrows with memory keys instead.