Fix display not clearing when full
Previously the logic to clear the display was nested inside a
conditional that triggered only when the length of the digits, n, to
display was less than the maximum number of digits in the display, N.
This prevented the display to clear after an operation with n >= N.

I discovered this issue while typing 3.8*3.8-9=5.44. Interestingly,
the display showed 5.440000, this comes from numerical imprecisions,
internally 3.8*3.8-9=5.4399999999999995, which after truncation to 8
digits and rounding, gives the displayed result. This motivated adding
an extra condition to remove trailing 0's after the decimal point.