Change operation logic
I found that allowing the use of prefixing and postfixing operators
lead to unexpected calculations, for example the sequence 'num'+'+'+
'±'+'=' displayed the result '-aa', 'sqrt'+'±'+'num'+'='' did not
return an error and similar bugs in cases when an unusual sequence of
operations was typed.

I decided to refactor the logic taking care of all the possible
combinations of operations, ruling out some unusual ones by showing
an error, for example a binary operation followed by a postfixing.

While testing the program I found some operations attempted to calculate
'ERROR' and displayed 'NAN'. I changed the logic to always display
'ERROR' and reset the display.