Fix delete behavior
Previously the delete key acted on the displayed digits and did not
removed the input from the display.current, causing strings such as
'num1' + 'delete' + 'num2' to display 'num1num2'. It also allowed the
user to erase letters from 'ERROR', instead of resetting the display.

A couple of other bugs related to prefixing and postfixing operators
were fixed, for example the sequence 'sqrt' + '±' didn't evaluate to
'ERROR'. After evaluating a postfixing operation the display didn't
reset if the user tried to input another number. The sequence
'±' + 'num' +'sqrt' didn't evaluate to error.

The current solution uses the property 'clear' of the display, which
was originally intended to be used for the solution that allowed
displaying large numbers. Probably a better solution can be done
splitting the 'unaryOperation' property into 'prefixing' and
'postfixing'.